segmentation/requirements.txt
absl-py==0.14.0
aiohttp==3.7.4.post0
amqp==2.6.1
arrow==0.14.7
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1620387926260/work
autobahn==21.2.1
Automat==20.2.0
backcall==0.1.0
billiard==3.6.3.0
bleach==3.1.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.2
celery==4.4.7
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625814693874/work
chardet==3.0.4
chart-studio==1.1.0
click==8.0.1
cloudpickle==1.6.0
codecov==2.1.0
configparser==5.0.2
constantly==15.1.0
coverage==5.3
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.22
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
dicom2nifti @ file:///home/conda/feedstock_root/build_artifacts/dicom2nifti_1622035074417/work
dill==0.3.3
distro-info==0.0.0
dlib==19.22.1
dm-tree==0.1.5
docker-pycreds==0.4.0
dominate==2.6.0
EasyProcess==0.2.8
entrypoints==0.3
estimator==0.0.10
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
Flask-Webpack==0.1.0
future==0.18.2
gast==0.3.3
gdcm==1.1
gdown==3.12.2
gevent==20.6.2
gitdb==4.0.7
GitPython==3.1.18
Glances==3.2.2
google-api-core==1.28.0
google-api-python-client==2.6.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
grpcio==1.40.0
gunicorn==20.0.4
h5py==3.1.0
httplib2==0.19.1
hyperlink==21.0.0
ics==0.7
idna==2.8
idna-ssl==1.1.0
imageio==2.9.0
imageio-ffmpeg==0.4.4
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata==4.8.1
incremental==21.3.0
ipydatawidgets==4.2.0
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
itk==5.2.0.post3
itk-core==5.2.0.post3
itk-filtering==5.2.0.post3
itk-io==5.2.0.post3
itk-numerics==5.2.0.post3
itk-registration==5.2.0.post3
itk-segmentation==5.2.0.post3
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==3.0.1
joblib==1.0.1
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
K3D==2.9.1
kaggle==1.5.10
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==4.6.11
lmdb==1.2.1
lpips==0.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
mpi4py==3.0.3
mudicom==0.1.2
multidict==5.1.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.5.1
neurite==0.1
ngl==1.6.1
nibabel==3.2.1
ninja==1.10.0.post2
notebook==6.0.2
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
olefile==0.46
opencv-python==4.0.1.24
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1625323647219/work
paho-mqtt==1.5.0
pam==0.2.0
pandas==1.1.4
pandocfilters==1.4.2
parse==1.12.1
parso==0.5.2
pathlib==1.0.1
pathtools==0.1.2
patool==1.12
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
plotly==5.3.1
pluggy==0.13.1
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==2.0.10
protobuf==3.18.0
psutil==5.7.3
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.5.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pydicom==2.2.1
pydocstyle==5.0.1
PyDrive==1.3.1
PyDrive2==1.8.3
pyflakes==2.1.1
Pygments==2.5.2
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1610910308735/work
pylama==7.7.1
pynrrd==0.4.2
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
pyrsistent==0.15.6
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1610291458349/work
pystrum==0.1
pytest==5.3.1
pytest-cov==2.8.1
pytest-html==2.0.1
pytest-metadata==1.8.0
pytest-mock==1.13.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-pam==1.8.4
python-slugify==4.0.1
pytictoc==1.5.0
pytorch-fid==0.2.0
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1633452062248/work
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
pyunpack==0.1.2
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.3.3
seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work
Send2Trash==1.5.0
sentry-sdk==1.4.3
setproctitle==1.2.2
shortuuid==1.0.1
SimpleITK==2.1.1
six==1.16.0
smmap==4.0.0
snowballstemmer==2.0.0
stringcase==1.2.0
subprocess32==3.5.4
Surface-Distance-Based-Measures @ file:///research/cbim/medical/qc58/work/projects/surface-distance
TatSu==4.4.0
tenacity==8.0.1
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1611075653546/work/tensorboard_plugin_wit-1.8.0-py3-none-any.whl
tensorflow==2.3.1
tensorflow-datasets==1.3.2
tensorflow-estimator==1.14.0
tensorflow-gan==2.0.0
tensorflow-gpu==1.14.0
tensorflow-hub==0.9.0
tensorflow-metadata==1.2.0
tensorflow-probability==0.11.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.2.0
tifffile==2020.9.3
torch==1.3.1
torchfile==0.1.0
torchinfo==1.5.3
torchvision==0.4.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1604105020994/work
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
traitlets==4.3.3
traittypes==0.2.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.25.11
utilspie==0.1.0
vine==1.3.0
visdom==0.1.8.9
voxelmorph==0.1
vtk==9.0.3
wandb==0.12.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wslink==1.0.7
yarl==1.6.3
yaspin==2.1.0
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
